requirements.txt
python-telegram-bot==21.3
google-generativeai==0.8.3
pymongo==4.8.0
